@@ -12,50 +12,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.4",
-    "@angular/cdk": "^10.0.2",
-    "@angular/common": "^10.0.4",
-    "@angular/compiler": "^10.0.4",
-    "@angular/core": "^10.0.4",
+    "@angular/animations": "^10.0.11",
+    "@angular/cdk": "^10.1.3",
+    "@angular/common": "^10.0.11",
+    "@angular/compiler": "^10.0.11",
+    "@angular/core": "^10.0.11",
     "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "^10.0.4",
-    "@angular/material": "^10.0.2",
-    "@angular/platform-browser": "^10.0.4",
-    "@angular/platform-browser-dynamic": "^10.0.4",
-    "@angular/router": "^10.0.4",
+    "@angular/forms": "^10.0.11",
+    "@angular/material": "^10.1.3",
+    "@angular/platform-browser": "^10.0.11",
+    "@angular/platform-browser-dynamic": "^10.0.11",
+    "@angular/router": "^10.0.11",
     "@swimlane/ngx-charts": "^14.0.0",
     "@waiting/base64": "^4.2.9",
-    "ajv": "^6.12.3",
+    "ajv": "^6.12.4",
     "angular-oauth2-oidc": "^10.0.3",
     "angular-oauth2-oidc-jwks": "^9.0.0",
-    "rxjs": "^6.6.0",
-    "rxjs-compat": "^6.6.0",
-    "tslib": "^1.13.0",
+    "csv-rxjs-kit": "^1.1.0",
+    "rxjs": "^6.6.2",
+    "rxjs-compat": "^6.6.2",
     "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1000.3",
-    "@angular/compiler-cli": "^10.0.4",
-    "@angular/language-service": "^10.0.4",
-    "@types/jasmine": "^3.5.11",
+    "@angular-devkit/build-angular": "^0.1000.6",
+    "@angular/compiler-cli": "^10.0.11",
+    "@angular/language-service": "^10.0.11",
+    "@types/jasmine": "^3.5.12",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^14.0.23",
-    "@typescript-eslint/eslint-plugin": "^3.6.1",
-    "@typescript-eslint/parser": "^3.6.1",
-    "eslint": "^7.5.0",
+    "@types/node": "^14.6.0",
+    "@typescript-eslint/eslint-plugin": "^3.9.1",
+    "@typescript-eslint/parser": "^3.9.1",
+    "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-no-null": "^1.0.2",
     "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.3",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "^5.0.2",
-    "karma": "^5.1.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "^3.0.3",
-    "karma-jasmine": "^3.3.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
+    "eslint-plugin-react": "^7.20.6",
     "prettier": "^2.0.5",
-    "protractor": "^7.0.0",
     "ts-node": "^8.10.2",
     "typescript": "^3.9.7"
   }
