@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable, of, zip, OperatorFunction, MonoTypeOperatorFunction } from 'rxjs';
-import { map, switchMap, toArray, skip, filter, takeWhile, startWith, endWith, tap } from 'rxjs/operators';
+import { Observable, of, zip, OperatorFunction } from 'rxjs';
+import { map, switchMap, toArray, filter, takeWhile } from 'rxjs/operators';
 
 import { autoMap, tuple, fromEntries } from 'src/app/utils/utils';
 
@@ -33,378 +33,9 @@ import {
   SDE_CSV_Types_Names_S,
 } from './models/eve-sde-types';
 
-/** Basic CSV-record.
- *
- * csvRecord datatype represents single MIME Type 'text/csv' [RFC 4180](https://www.ietf.org/rfc/rfc4180.txt) data record, including header.
- * Any first record of a CSV-record stream might be header record.
- *
- * Every CSV file record just an array of string. Empty records are empty arrays. Every empty field is an empty string.
- */
-type csvRecord = string[];
-
-/**
- * CSV-record validator function type.
- * @param csvRecord CSV-record to validate
- * @param isHeader CSV-record type
- * @returns validated CSV-record
- * @throws any validation errors
- */
-type csvRecordValidator = (rec: csvRecord, isHeader: boolean) => csvRecord;
-
-/**
- * Object builder function type.
- * @template T target object type
- * @param rec source CSV-record
- * @param names fields names for the record (if available)
- * @returns object
- */
-type csvObjectBuilder<T> = (rec: csvRecord, names?: string[]) => T;
-
-/**
- * Field names extractor funcion type.
- * @template T object type
- * @param obj object to exctract names
- * @returns field names
- */
-type csvHeaderExtractor<T> = (obj: T) => string[];
-
-/**
- * CSV-record extractor function type.
- * @template T object type
- * @param obj object to extract CSV-record data
- * @param names field names array (if available)
- * @returns CSV-record
- */
-type csvRecordExtractor<T> = (obj: T, names?: string[]) => csvRecord;
-
-/**
- * CSV formatting output.
- *
- * RxJS operator converts CSV-record stream to MIME Type 'text/csv' [RFC 4180](https://www.ietf.org/rfc/rfc4180.txt) text stream.
- *
- * The operator changes stream type from 'csvRecord' to 'string'.
- *
- * @param opt Options to format output.
- */
-function csvStringify(opt?: { eol?: '\r\n' | 'n'; eof?: boolean }): OperatorFunction<csvRecord, string> {
-  const eol = opt?.eol || '\r\n';
-  const line$: OperatorFunction<csvRecord, string> = map(
-    (r, i) =>
-      (i ? eol : '') + r.map((field) => (/[,"\r\n]/.test(field) ? `"${field.replace(/"/g, '""')}"` : field)).join(',')
-  );
-  return opt?.eof ? (data$: Observable<csvRecord>) => data$.pipe(line$, endWith(eol)) : line$;
-}
-
-/**
- * CSV formattion input.
- *
- * RxJS operator converts MIME Type 'text/csv' [RFC 4180](https://www.ietf.org/rfc/rfc4180.txt)  text stream to CSV-record stream.
- *
- * The operator changes stream type from 'string' to 'csvRecord'.
- */
-function csvParse(): OperatorFunction<string, csvRecord> {
-  return function (csv$: Observable<string>): Observable<csvRecord> {
-    return new Observable((observer) => {
-      enum S {
-        IDLE,
-        COMMA,
-        CR,
-        TEXT,
-        ESCAPED,
-        CLOSED,
-        ERROR,
-      }
-      const R = new RegExp(/,|"|\n|\r\n|\r|[^,"\r\n]+/y);
-      let state: S = S.IDLE;
-      let r: csvRecord = [];
-      let v = '';
-      let rIndex = 1;
-      let cIndex = 1;
-      let lastToken = '';
-      function err(msg: string): void {
-        state = S.ERROR;
-        observer.error(new SyntaxError(`CSV [${rIndex}, ${cIndex}]: ` + msg));
-      }
-      function next(): void {
-        observer.next(r);
-        r = [];
-      }
-      function fsm(s: string): boolean {
-        switch (state) {
-          case S.IDLE: // start of a line
-          case S.CR: // right after <CR>
-            switch (s) {
-              case ',':
-                r.push('');
-                state = S.COMMA;
-                break;
-              case '"':
-                state = S.ESCAPED;
-                break;
-              case '\r':
-              case '\r\n':
-                next();
-                state = s === '\r' ? S.CR : S.IDLE;
-                break;
-              case '\n':
-                if (state === S.CR) state = S.IDLE;
-                else next();
-                break;
-              case '':
-                break;
-              default:
-                v += s;
-                state = S.TEXT;
-                break;
-            }
-            break;
-          case S.COMMA: // right after ','
-            switch (s) {
-              case ',':
-                r.push('');
-                break;
-              case '"':
-                state = S.ESCAPED;
-                break;
-              case '\r':
-              case '\r\n':
-              case '\n':
-              case '':
-                r.push('');
-                next();
-                state = s === '\r' ? S.CR : S.IDLE;
-                break;
-              default:
-                v += s;
-                state = S.TEXT;
-                break;
-            }
-            break;
-          case S.TEXT: // inside non-escaped text
-            switch (s) {
-              case ',':
-                r.push(v);
-                v = '';
-                state = S.COMMA;
-                break;
-              case '"':
-                err('Quote in non-escaped data.');
-                return false;
-              case '\r':
-              case '\r\n':
-              case '\n':
-              case '':
-                r.push(v);
-                v = '';
-                next();
-                state = s === '\r' ? S.CR : S.IDLE;
-                break;
-              default:
-                v += s;
-                break;
-            }
-            break;
-          case S.ESCAPED: // inside escaped text
-            if (s === '"') state = S.CLOSED;
-            else if (s === '') {
-              err('Closing quote is missing.');
-              return false;
-            } else v += s;
-            break;
-          case S.CLOSED: // after '"' inside escaped text
-            switch (s) {
-              case ',':
-                r.push(v);
-                v = '';
-                state = S.COMMA;
-                break;
-              case '"':
-                v += s;
-                state = S.ESCAPED;
-                break;
-              case '\r':
-              case '\r\n':
-              case '\n':
-              case '':
-                r.push(v);
-                v = '';
-                next();
-                state = s === '\r' ? S.CR : S.IDLE;
-                break;
-              default:
-                err('Invalid escape sequence.');
-                return false;
-            }
-            break;
-          default:
-            return false;
-        }
-        if (s[0] === '\r' || (s === '\n' && lastToken !== '\r')) {
-          rIndex++;
-          cIndex = 1;
-        } else cIndex += s.length;
-        lastToken = s;
-        return true;
-      }
-      return csv$.subscribe({
-        next(chunk: string): void {
-          if (state === S.ERROR) return;
-          let m: string | undefined;
-          while ((m = R.exec(chunk)?.[0]) != undefined && fsm(m));
-        },
-        complete(): void {
-          if (state === S.ERROR) return;
-          if (fsm('')) observer.complete();
-        },
-        error(e: unknown): void {
-          if (state === S.ERROR) return;
-          observer.error(e);
-        },
-      });
-    });
-  };
-}
-
-/**
- * Removes empty records.
- *
- * RxJS operator removes all empty records from CSV-record stream.
- *
- * The operator doesn't change stream type.
- */
-function csvDropEmpty(): MonoTypeOperatorFunction<csvRecord> {
-  return filter((r) => !!r.length);
-}
-
-/**
- * Removes the header.
- *
- * RxJS operator removes the header form CSV-record stream. It just removes first record form a stream. There is no dedicated tagging
- * for header record so one must be sure the stream has a header before removing it.
- *
- * The operator removes additional 'csvRecord' type from a stream and leaves only the main type T, which might be 'csvRecord'.
- * @template T the main stream type, might be csvRecord or any arbitrary data type
- */
-function csvDropHeader<T>(): OperatorFunction<T | csvRecord, T> {
-  return skip(1) as OperatorFunction<T | csvRecord, T>;
-}
-
-/**
- * Injects a header into a data stream.
- *
- * RxJS operator inserts new record into the stream at first position.
- * Inserted record might be interpreted like a header by other operators if they are instructed to.
- *
- * The operator adds 'csvRecord' type to the main stream type.
- * @template T the main stream type
- * @param header header record value
- */
-function csvInjectHeader<T>(header: csvRecord): OperatorFunction<T, T | csvRecord> {
-  return startWith(header);
-}
-
-function csvValidateRecord(hdr: boolean, validator: csvRecordValidator): MonoTypeOperatorFunction<csvRecord> {
-  return map((r, i) => validator(r, hdr && !i));
-}
-
-function csvRebuildObject(hdr: false): OperatorFunction<csvRecord, string[]>;
-function csvRebuildObject(hdr: true): OperatorFunction<csvRecord, string[] | csvRecord>;
-//function csvRebuildObject(hdr: 'drop'): OperatorFunction<csvRecord, string[]>;
-function csvRebuildObject<T>(hdr: false, prj: csvObjectBuilder<T>): OperatorFunction<csvRecord, T>;
-function csvRebuildObject<T>(hdr: true, prj: csvObjectBuilder<T>): OperatorFunction<csvRecord, T | csvRecord>;
-//function csvRebuildObject<T>(hdr: 'drop', prj: csvObjectBuilder<T>): OperatorFunction<csvRecord, T>;
-function csvRebuildObject<T>(
-  hdr: false | true /*| 'drop'*/,
-  prj?: csvObjectBuilder<T>
-):
-  | OperatorFunction<csvRecord, string[]>
-  | OperatorFunction<csvRecord, string[] | csvRecord>
-  | OperatorFunction<csvRecord, T>
-  | OperatorFunction<csvRecord, T | csvRecord> {
-  if (!prj) return /*hdr === 'drop' ? csvDropHeader<T>() :*/ (o) => o;
-  if (!hdr) return map((rec) => prj(rec));
-  return (data$: Observable<csvRecord>) => {
-    return new Observable<csvRecord | T>((observer) => {
-      let names: string[] | undefined = undefined;
-      return data$
-        .pipe(
-          map((obj, i) => {
-            if (i !== 0) return prj(obj, names);
-            names = obj as string[];
-            return obj;
-          })
-          //filter((_, i) => i !== 0 || hdr !== 'drop')
-        )
-        .subscribe(observer);
-    });
-  };
-}
-
-//function csvExtractRecord(hdr: string[]): OperatorFunction<unknown[], csvRecord>;
-function csvExtractRecord(hdr: false): OperatorFunction<unknown[], csvRecord>;
-function csvExtractRecord(hdr: true): OperatorFunction<csvRecord | unknown[], csvRecord>;
-//function csvExtractRecord<T>(hdr: string[], prj: csvRecordExtractor<T>): OperatorFunction<T, csvRecord>;
-function csvExtractRecord<T>(hdr: false, prj: csvRecordExtractor<T>): OperatorFunction<T, csvRecord>;
-function csvExtractRecord<T>(hdr: true, prj: csvRecordExtractor<T>): OperatorFunction<csvRecord | T, csvRecord>;
-function csvExtractRecord<T>(
-  hdr: /*string[] |*/ false | true,
-  prj?: csvRecordExtractor<T>
-):
-  | OperatorFunction<unknown[], csvRecord>
-  | OperatorFunction<T, csvRecord>
-  | OperatorFunction<csvRecord | unknown[], csvRecord>
-  | OperatorFunction<csvRecord | T, csvRecord> {
-  const u2s = (obj: unknown[]) => obj.map((v) => String(v ?? ''));
-  if (!hdr) return prj ? map<T, csvRecord>((obj) => prj(obj)) : map<unknown[], csvRecord>((obj) => u2s(obj));
-  return (data$: Observable<unknown[]> | Observable<T>) =>
-    new Observable<csvRecord>((observer) => {
-      let names: string[] | undefined = /*Array.isArray(hdr) ? hdr :*/ undefined;
-      return (data$ as Observable<unknown>)
-        .pipe(
-          map((obj, i) => {
-            if (i !== 0 /*|| names*/) return prj ? prj(obj as T, names) : u2s(obj as unknown[]);
-            names = obj as string[];
-            return obj;
-          }) as OperatorFunction<unknown, csvRecord>
-        )
-        .subscribe(observer);
-    });
-}
-
-function csvRecordsJustifier(opt?: { hdr_length?: number; fill_empty?: true; filler?: string }): csvRecordValidator {
-  let hlen = opt?.hdr_length;
-  return (rec: csvRecord, isHeader: boolean) => {
-    const rlen = rec.length;
-    if (isHeader) hlen = rlen;
-    else if (hlen && (rlen != 0 || opt?.fill_empty)) {
-      if (rlen < hlen) rec = rec.concat(Array(hlen - rlen).fill(opt?.filler ?? ''));
-      if (rlen > hlen) rec = rec.slice(0, hlen);
-    }
-    return rec;
-  };
-}
-
-function csvObjAssembler(extra?: string): csvObjectBuilder<Record<string, string>> {
-  const n = extra || '';
-  return (rec: csvRecord, names?: string[]) => {
-    const fields = names || [];
-    return Object.assign({}, ...rec.map((v, i) => ({ [fields[i] || `${n}#${i}`]: v }))) as Record<string, string>;
-  };
-}
-
-function csvObjPropsGetter(): csvHeaderExtractor<Record<string, unknown>> {
-  return (obj: Record<string, unknown>) => Object.keys(obj).sort((a, b) => a.localeCompare(b));
-}
-
-function csvObjValuesGetter(extra?: boolean): csvRecordExtractor<Record<string, unknown>> {
-  return (obj: Record<string, unknown>, names?: string[]) => {
-    let listed = names || [];
-    if (extra) listed = listed.concat(csvObjPropsGetter()(obj).filter((k) => !listed.includes(k)));
-    return listed.map((k) => String(obj[k] ?? ''));
-  };
-}
-
 import Ajv from 'ajv';
+
+import { csvParse, csvConvert, csvAssembler, csvDropHeader } from 'csv-rxjs-kit';
 
 class ValidationError extends Error {
   public readonly name = 'ValidationError';
@@ -417,7 +48,7 @@ class ValidationError extends Error {
 function jsonValidate<T>(
   schema: Record<string, unknown>,
   opt?: { coerce?: boolean },
-  completeOnError?: boolean | ((e: ValidationError) => (boolean | unknown))
+  completeOnError?: boolean | ((e: ValidationError) => boolean | unknown)
 ): OperatorFunction<unknown, T> {
   return (data$: Observable<unknown>) =>
     new Observable((observer) => {
@@ -453,7 +84,7 @@ export class SdeService {
       .get(SDE_BASE + name, { headers: { Accept: 'text/csv; header=present' }, responseType: 'text' })
       .pipe(
         csvParse(),
-        csvRebuildObject(true, csvObjAssembler()),
+        csvConvert(true, csvAssembler()),
         csvDropHeader(),
         jsonValidate<T>(schema, { coerce: true }, false),
         toArray()
